dummy_devices/requirements.txt
paho-mqtt==1.5.1
opencv-python==4.9.0.80
numpy